Fix emptyview and pertag conflict
The pertag patch was not focusing or arranging clients if the newtagset was
empty. This would be fine for keybindings that called the toggleview function,
but calling the toggleview function in any other case where the clients weren't
being focused or rearranged would cause the new tagset to take effect, but the
clients would still be shown until hovering over it would cause a focus or
arrange call.

This in particular was causing issues with the polybar-dwm-module which would
change the tagset to an empty view, but cause the clients to remain visible,
until they were hovered over.

There is also an additional check for the curtag, prevtag assignment that checks
if newtagset is empty, otherwise the for loop would loop forever.